Create Code
This document describes how to produce a set of graphics and perform the associated statistical tests that describe trends in temperature at two shallow lakes. 
We are interested in depicting (graphically and statistically) how surface temperatures vary at two distinct shallow lakes. Lakes are located within the same watershed and equipped with the same type of HOBO datalogger ($Onset^{TM}$) tethered to a tree in the littoral zone of the shallow lake (distance to shore < 0.5 meter), and let floating at the surface. Dataloggers thus measure temperature at the interface between the water column and the air compartment. The two dataloggers (named Arj.20125800.S and Arj.20125804.N) record continuous data with a timelag of 30 minutes each. Given that local conditions (vegetation, shade ...) and microclimate may differ between the two shallow lakes, we aim at describing the trends and also assessing whether the two timeseries differ or not.